+++ symbolic/src/main/kotlin/mist/module/ModuleSymbol.kt
@@ -2,6 +2,6 @@ package mist.module
 
 data class ModuleSymbol(
   val name: String,
-  val address: Long, // TODO int?
+  val address: Int,
   val length: Int
 )
